fix(fileio): S3.7-F4 — fs_truncate unknown-fd format byte-drift
Wave-3 S3.7 review found a new byte-drift in fs_truncate's
`unknown fd` error message.  Pre-refactor destructured `fd: u64`
from the parsed tuple:

    let parsed = match ... { (Some(fd), Some(n)) if n >= 0 =>
                             Some((fd as u64, n as u64)), ... };
    match parsed {
        Some((fd, n)) => {                           // fd: u64
            ...
            None => err(FSERR_CLOSED, format!("unknown fd {fd}")),
        }
    }

The format thus uses UNSIGNED formatting.  My S3.7 migration
"corrected" via `args.fd as i64` — but that cast came from the
S3.6 fs_seek pattern, where pre-refactor destructured `fd: i64`
directly from `RhoNumber::unapply`.  fs_seek's signed format is
correct to preserve; fs_truncate's is UNSIGNED.

Impact under determinism: none (leader and follower agree on the
format).  But the T-04 byte-identity claim in S3.7's commit
message is technically false for this specific error path.
Reverting `args.fd as i64` → `args.fd` in fs_truncate restores
pre-refactor byte-identity.

fs_seek's `as i64` cast (S3.6) STAYS — that one matches
pre-refactor.  The pre-refactor codebase is intentionally
inconsistent between fs_seek (i64 → signed) and fs_truncate (u64
→ unsigned); wave 3 must preserve that inconsistency.

Verification:
- cargo build --lib -p rholang: clean.
- cargo test -p rholang --test fs_wal_spec: 110 passed.

The historical F2 drift in fs_flush / fs_tell / fs_size (all
using `format!("unknown fd {}", args.fd)` where args.fd is u64
but pre-refactor `fd` was i64) remains user-deferred.

+++ rholang/src/rust/interpreter/io/handlers.rs
@@ -6513,11 +6513,14 @@ impl FsHandler for FsTruncateHandler {
             }
             let file_arc = match ctx.handles.raw_fd(args.fd).await {
                 Some(f) => f,
+                // Pre-refactor fs_truncate destructured `fd: u64` from
+                // the parsed tuple `Some((fd as u64, n as u64))`, so
+                // the format is UNSIGNED.  Different from fs_seek /
+                // fs_size where fd is `i64` from RhoNumber::unapply
+                // and format is signed.  Preserve pre-refactor
+                // byte-identity — no `as i64` cast here.
                 None => {
-                    return HandlerReply::err(
-                        FSERR_CLOSED,
-                        format!("unknown fd {}", args.fd as i64),
-                    );
+                    return HandlerReply::err(FSERR_CLOSED, format!("unknown fd {}", args.fd));
                 }
             };
             let n = args.n;
